modify SMO SDK samples
