--- doc/smocodesample/net40app/app.config
@@ -1,23 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.SqlServer.BatchParser" publicKeyToken="89845dcd8080cc91" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-13.100.0.0" newVersion="13.100.0.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.SqlServer.ConnectionInfo" publicKeyToken="89845dcd8080cc91" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-13.100.0.0" newVersion="13.100.0.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.SqlServer.Diagnostics.STrace" publicKeyToken="89845dcd8080cc91" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-13.100.0.0" newVersion="13.100.0.0" />
-      </dependentAssembly>
-      <dependentAssembly>
-        <assemblyIdentity name="Microsoft.SqlServer.SString" publicKeyToken="89845dcd8080cc91" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-13.100.0.0" newVersion="13.100.0.0" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-</configuration>
