--- doc/smocodesample/net40app/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="SharedManagementObjects" version="140.1.9" targetFramework="net40" />
-</packages>
